Added redirection to Makefile
