+++ Makefile
@@ -19,11 +19,13 @@ ifeq ($(O_SYSTEM),$(filter $(O_SYSTEM),MacOS Linux))
 	EX_MANIP_EXEC := manipulators
 	EX_PB_EXEC := progressbar
 	EX_GRAPH_EXEC := graphics
+	EX_REDIR_EXEC := redirection
 	TEST_EXEC := tests
 else
 	EX_MANIP_EXEC := manipulators.exe
 	EX_PB_EXEC := progressbar.exe
 	EX_GRAPH_EXEC := graphics.exe
+	EX_REDIR_EXEC := redirection.exe
 	TEST_EXEC := tests.exe
 endif
 LIB := libosmanip.a
@@ -50,6 +52,7 @@ ifeq ($(O_SYSTEM),$(filter $(O_SYSTEM),MacOS Linux))
 	SRC_MANIP := $(shell find $(SRC_DIR) -name '*.cpp') $(shell find $(EX_DIR) -name 'manipulators.cpp')
 	SRC_PB := $(shell find $(SRC_DIR) -name '*.cpp') $(shell find $(EX_DIR) -name 'progressbar.cpp')
 	SRC_GRAPH := $(shell find $(SRC_DIR) -name '*.cpp') $(shell find $(EX_DIR) -name 'graphics.cpp')
+	SRC_REDIR := $(shell find $(SRC_DIR) -name '*.cpp') $(shell find $(EX_DIR) -name 'redirection.cpp')
 
 	SRC_LIB := $(shell find $(SRC_DIR) -name '*.cpp')
 	TEST := $(shell find $(SRC_DIR) -name '*.cpp') $(shell find $(TEST_DIR) -name '*.cpp')
@@ -58,6 +61,7 @@ else
 	SRC_MANIP := $(ALL_SOURCES) $(wildcard $(EX_DIR)/manipulators.cpp)
 	SRC_PB := $(ALL_SOURCES) $(wildcard $(EX_DIR)/progressbar.cpp)
 	SRC_GRAPH := $(ALL_SOURCES) $(wildcard $(EX_DIR)/graphics.cpp)
+	SRC_REDIR := $(ALL_SOURCES) $(wildcard $(EX_DIR)/redirection.cpp)
 
 	SRC_LIB := $(ALL_SOURCES)
 	ALL_TESTS := $(wildcard $(TEST_DIR)/*.cpp) $(wildcard $(TEST_DIR)/*/*.cpp)
@@ -72,6 +76,7 @@ endif
 OBJ_MANIP := $(SRC_MANIP:%=$(OBJ_DIR)/%.o)
 OBJ_PB := $(SRC_PB:%=$(OBJ_DIR)/%.o)
 OBJ_GRAPH := $(SRC_GRAPH:%=$(OBJ_DIR)/%.o)
+OBJ_REDIR := $(SRC_REDIR:%=$(OBJ_DIR)/%.o)
 
 # Other source objects
 OBJ_LIB := $(SRC_LIB:%=$(OBJ_DIR)/%.o)
@@ -80,7 +85,7 @@ TEST_OBJ := $(TEST:%=$(OBJ_DIR)/%.o)
 #====================================================
 #     Dependencies and flags
 #====================================================
-DEPS := $(OBJ_MANIP:.o=.d) $(OBJ_PB:.o=.d) $(OBJ_GRAPH:.o=.d)
+DEPS := $(OBJ_MANIP:.o=.d) $(OBJ_PB:.o=.d) $(OBJ_GRAPH:.o=.d $(OBJ_REDIR:.o=.d)
 ifeq ($(O_SYSTEM),$(filter $(O_SYSTEM),MacOS Linux))
 	INC_DIR := $(shell find $(SRC_DIR) -type d)
 else
@@ -97,8 +102,8 @@ LDFLAGS := -pthread -larsenalgear
 #====================================================
 .PHONY: clean all
 
-all: $(BUILD_DIR)/$(EX_MANIP_EXEC) $(BUILD_DIR)/$(EX_PB_EXEC) $(BUILD_DIR)/$(EX_GRAPH_EXEC) $(BUILD_DIR)/$(TEST_EXEC) $(LIB_DIR)/$(LIB)
-examples: $(BUILD_DIR)/$(EX_MANIP_EXEC) $(BUILD_DIR)/$(EX_PB_EXEC) $(BUILD_DIR)/$(EX_GRAPH_EXEC) $(LIB_DIR)/$(LIB)
+all: $(BUILD_DIR)/$(EX_MANIP_EXEC) $(BUILD_DIR)/$(EX_PB_EXEC) $(BUILD_DIR)/$(EX_GRAPH_EXEC) $(BUILD_DIR)/$(EX_REDIR_EXEC) $(BUILD_DIR)/$(TEST_EXEC) $(LIB_DIR)/$(LIB)
+examples: $(BUILD_DIR)/$(EX_MANIP_EXEC) $(BUILD_DIR)/$(EX_PB_EXEC) $(BUILD_DIR)/$(EX_GRAPH_EXEC) $(BUILD_DIR)/$(EX_REDIR_EXEC) $(LIB_DIR)/$(LIB)
 tests: $(BUILD_DIR)/$(TEST_EXEC)
 
 #====================================================
@@ -119,6 +124,11 @@ $(BUILD_DIR)/$(EX_PB_EXEC): $(OBJ_PB)
 $(BUILD_DIR)/$(EX_GRAPH_EXEC): $(OBJ_GRAPH)
 	@ mkdir -p $(dir $@)
 	$(CXX) $(OBJ_GRAPH) -o $@ $(LDFLAGS)
+
+# Redirection
+$(BUILD_DIR)/$(EX_REDIR_EXEC): $(OBJ_REDIR)
+	@ mkdir -p $(dir $@)
+	$(CXX) $(OBJ_REDIR) -o $@ $(LDFLAGS)
 
 #====================================================
 #     Building tests
